parent c04782a629af329b10c3d7a8604f47eea52ae7df author Hadar Sharon <hsharon@dhcp-4-126.tlv.redhat.com> 1551691943 +0200 committer Hadar Sharon <hsharon@localhost.localdomain> 1553765953 +0200
Removing now-ignored files and dirs

Remove Cockpit

It's not used and not useful right now.

Added node-4

Create glusto YAML config as part of the Vagrantfile invocation.

Not a complete solution, but seems reasonable:
create the Glusto basic YAML file as part of the Vagrantfile.

In addition, try to cleanup the 'testvol' created.
The volume is removed - next phase would be to remove the bricks
and the mounts and the LVs, let Glusto take over from scratch.

Create glusto YAML config as part of the Vagrantfile invocation.

Not a complete solution, but seems reasonable:
create the Glusto basic YAML file as part of the Vagrantfile.

In addition, try to cleanup the 'testvol' created.
The volume is removed - next phase would be to remove the bricks
and the mounts and the LVs, let Glusto take over from scratch.

Modified gluster-cleanup.yml to remove existing volumes before testing, added glusterfs to node-0