Create glusto YAML config as part of the Vagrantfile invocation.
Not a complete solution, but seems reasonable:
create the Glusto basic YAML file as part of the Vagrantfile.

In addition, try to cleanup the 'testvol' created.
The volume is removed - next phase would be to remove the bricks
and the mounts and the LVs, let Glusto take over from scratch.